clients/python/pyproject.toml
[project]
name = "clients"
version = "0.1.0"
description = "Internal shared clients"
requires-python = ">=3.12"
dependencies = [
    "couchbase>=4.4.0",
    "pydantic>=2.0.0",
    "httpx>=0.27.0",
    "resonate-sdk>=0.6.7",
]


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
